requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Agent Framework & LLMs
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10
transformers>=4.35.0
torch==2.1.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# Mistral AI (choose one based on usage)
# Option 1: Mistral API SDK
mistralai>=0.1.0
# Option 2: HuggingFace transformers (for local models)
# Already included above

# Speech Recognition
openai-whisper>=20231117
# Alternative: faster-whisper for optimized performance
# faster-whisper>=0.9.0

# Audio Processing
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0

# Vector Database
faiss-cpu>=1.7.4
# For GPU acceleration, use: faiss-gpu>=1.7.4
sentence-transformers>=2.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
tenacity>=8.2.0
loguru>=0.7.0
python-dateutil>=2.8.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Optional: Speaker Diarization
pyannote.audio==3.1.1  # Speaker diarization (stable version, compatible with torch 2.1.0)
torchaudio==2.1.0  # Required for pyannote.audio 3.1.1 (matches torch version)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=19.0.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0
pylint>=3.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Development Tools
ipython>=8.17.0
jupyter>=1.0.0

